--- app/src/main/java/com/techcourse/support/web/filter/CharacterEncodingFilter.java
@@ -1,28 +1,0 @@
-package com.techcourse.support.web.filter;
-
-import jakarta.servlet.*;
-import jakarta.servlet.annotation.WebFilter;
-
-import java.io.IOException;
-
-@WebFilter("/*")
-public class CharacterEncodingFilter implements Filter {
-
-    private static final String DEFAULT_ENCODING = "UTF-8";
-
-    @Override
-    public void init(final FilterConfig filterConfig) throws ServletException {
-    }
-
-    @Override
-    public void doFilter(final ServletRequest request, final ServletResponse response, FilterChain chain)
-            throws IOException, ServletException {
-        request.setCharacterEncoding(DEFAULT_ENCODING);
-        response.setCharacterEncoding(DEFAULT_ENCODING);
-        chain.doFilter(request, response);
-    }
-
-    @Override
-    public void destroy() {
-    }
-}
